New feature (Transmit Time) and bug fixing
Add functionality:
Set a transmit time (in the system menu).
When you transmit, and you pass the 80% of the  set transmit time, the shown timer on the display turns red.
You can transmit past the time you have set. It's only informative.

Bug correction:
When exiting the System Menu the display was showing the set power level in big character.
When exiting the System Menu the code that was handling the rotation of the first rotary encoder was activated, changing the value of the paramater assinged to the encoder.